+++ src/app/src/actions/filters.js
@@ -13,6 +13,7 @@ export const updateContributorTypeFilter = createAction('UPDATE_CONTRIBUTOR_TYPE
 export const updateCountryFilter = createAction('UPDATE_COUNTRY_FILTER');
 export const updateCombineContributorsFilterOption = createAction('UPDATE_COMBINE_CONTRIBUTORS_FILTER_OPTION');
 export const updateBoundaryFilter = createAction('UPDATE_BOUNDARY_FILTER');
+export const updatePPEFilter = createAction('UPDATE_PPE_FILTER');
 export const resetAllFilters = createAction('RESET_ALL_FILTERS');
 export const updateAllFilters = createAction('UPDATE_ALL_FILTERS');