+++ src/django/api/views.py
@@ -494,7 +494,7 @@ def log_download(request):
 class FacilitiesAPIFilterBackend(BaseFilterBackend):
     def get_schema_fields(self, view):
         if view.action == 'list':
-            return [
+            fields = [
                 coreapi.Field(
                     name='q',
                     location='query',
@@ -550,6 +550,23 @@ class FacilitiesAPIFilterBackend(BaseFilterBackend):
                         'facilities within the boundaries of that geometry.')
                 ),
             ]
+
+            if switch_is_active('ppe'):
+                fields.append(
+                    coreapi.Field(
+                        name='ppe',
+                        location='query',
+                        type='boolean',
+                        required=False,
+                        description=(
+                            'If "true" only facilities with PPE '
+                            'production details or PPE-specific contact '
+                            'information will be returned. Any other value '
+                            'will return all facilities.'),
+                    )
+                )
+
+            return fields
 
         if view.action == 'create':
             return [